* moves accordion component into separate folder
