+++ packages/ui-library/src/components/collapsible/accordion/accordion.test.tsx
@@ -1,31 +1,8 @@
 import {describe, expect, test} from 'vitest';
 import {render} from '@testing-library/react';
 import {Accordion} from './accordion';
-import {Details} from './details';
+import {Details} from '../details';
 import React from 'react';
-
-describe('details', () => {
-  test('should render details', () => {
-    const {container} = render(<Details summary={'Testing'}>Hello World</Details>);
-
-    const detail = container.getElementsByClassName('details');
-    expect(detail).toBeDefined();
-  });
-
-  test('should render details in dark mode', () => {
-    const {container} = render(<Details summary={'Testing'} dark>Hello World</Details>);
-
-    const detail = container.getElementsByClassName('details dark');
-    expect(detail).toBeDefined();
-  });
-
-  test('should render details with divider', () => {
-    const {container} = render(<Details summary={'Testing'} divider>Hello World</Details>);
-
-    const detail = container.getElementsByClassName('details divider');
-    expect(detail).toBeDefined();
-  });
-});
 
 describe('accordion', () => {
   test('should render accordion', () => {
@@ -35,7 +12,7 @@ describe('accordion', () => {
         <Details summary={'test'}>Hello world</Details>
       </Accordion>
     );
-    
+
     const accordion = container.getElementsByClassName('accordion')[0];
     expect(accordion).toBeDefined();
     expect(accordion.children.length).toEqual(2);
