+++ packages/ui-library/src/components/collapsible/accordion/index.ts
@@ -1,2 +1,1 @@
-export * from './details';
 export * from './accordion';
