+++ packages/ui-library/src/components/collapsible/collapsible.module.scss
