+++ packages/ui-library/src/components/collapsible/details/details.stories.tsx
