+++ packages/ui-library/src/components/collapsible/details/index.ts
@@ -1,0 +1,1 @@
+export * from './details';
