+++ packages/ui-library/src/components/collapsible/details/types.ts
@@ -1,0 +1,10 @@
+import type {BaseComponentProps} from '../../common/types';
+import type {ReactNode} from 'react';
+
+export type DetailsProps = BaseComponentProps<'details'> & {
+  children: ReactNode;
+  summary: ReactNode;
+  divider?: boolean;
+  icon?: ReactNode;
+  iconPosition?: 'start' | 'end' | undefined;
+}
