+++ packages/ui-library/src/index.ts
@@ -4,7 +4,8 @@ export {RadioGroup, type RadioOption} from './components/check/radioGroup';
 export {CircularProgress, LinearProgress} from './components/progress';
 export {Pagination, type AriaLabels} from './components/pagination';
 export {Input, InputDecorator, Textarea} from './components/input';
-export {Details, Accordion} from './components/details';
+export {Accordion} from './components/collapsible/accordion';
+export {Details} from './components/collapsible/details';
 export {Checkbox} from './components/check/checkbox';
 export {Breadcrumbs} from './components/breadcrumbs';
 export {Radio} from './components/check/radio';
